final_project/app/faust_app/requirements.txt
confluent-kafka
faust-streaming
fastavro
httpx
cachetools
authlib
avro